requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
python-multipart
aiofiles
jinja2
lxml
